+++ CHANGELOG.md
@@ -3,8 +3,7 @@
 * The t method accepts multiple keys and options
 * Better specs
 * Unimplemented i18next methods raise an exception
-* Support methods exists, languages, loadNamespaces and resolvedLanguage
-* Add .yardopts to the gem file to show documentation at rubydoc.info
+* Support methods addResource, exists, getResource, languages, loadNamespaces and resolvedLanguage
 * Specs use a local copy of the i18next-fetch-backend plugin
 
 ## 0.1.3 - 2022-08-24